Feature - added new routes/component - removed previous components
