+++ src/router/routes.js
@@ -5,7 +5,7 @@ const routes = [
     component: () => import('layouts/MainLayout.vue'),
     children: [
       // { path: '', component: () => import('pages/Index.vue') }
-      { path: '', component: () => import('pages/Authentication/Login.vue') },
+      { path: '', component: () => import('pages/Authentication/Authentication.vue') },
     ]
   },
